Consolidate lsp
