+++ .vscode/launch.json
@@ -26,7 +26,7 @@
 				"script": "watch"
 			},
 			"env": {
-				"SERVER_PATH": "${workspaceFolder}/target/debug/queryscript-lsp"
+				"SERVER_PATH": "${workspaceFolder}/target/debug/qs-lsp"
 			},
 		},
 		{
@@ -44,7 +44,7 @@
 			],
 			"autoAttachChildProcesses": true,
 			"env": {
-				"SERVER_PATH": "${workspaceRoot}/target/debug/queryscript-lsp"
+				"SERVER_PATH": "${workspaceRoot}/target/debug/qs-lsp"
 			},
 		}
 	],
