--- lsp/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "queryscript-lsp"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tower-lsp = "0.17"
-tokio = { version = "^1.23", features = [
-	"io-util",
-	"io-std",
-	"macros",
-	"rt-multi-thread",
-	"fs",
-] }
-regex = "1"
-serde_json = "1"
-serde = { version = "*" }
-
-queryscript = { path = "../queryscript", features = ["multi-thread", "serde"] }
