+++ queryscript/Cargo.toml
@@ -22,12 +22,18 @@ name = "qs"
 path = "bin/qs.rs"
 required-features = ["cli"]
 
+[[bin]]
+name = "qs-lsp"
+path = "bin/lsp.rs"
+required-features = ["lsp"]
+
 [features]
 default = ["cli"]
 multi-thread = ["tokio/rt-multi-thread"]
 serde = ["dep:serde", "arrow-schema/serde", "half/serde"]
 ts = ["serde", "dep:ts-rs"]
 cli = ["clap", "home", "rustyline"]
+lsp = ["multi-thread", "tokio/io-util", "tokio/io-std", "tokio/macros", "tokio/fs", "regex", "serde_json", "serde", "tower-lsp"]
 
 [dependencies]
 
@@ -54,11 +60,14 @@ futures = "0.3"
 half = { version = "2.1" }
 lazy_static = "1.4.0"
 object_store = "0.5.0"
+regex = { version = "1", optional = true}
 serde = { version = "1", optional = true }
+serde_json = { version = "1", optional = true }
 snafu = { version = "0.7.3" }
 sqlparser = { path = "../sqlparser-rs", version = "0.30.0", package = "sqlparser-queryscript" }
 tabled = "0.10"
 tokio = "1.25"
+tower-lsp = { version = "0.17", optional = true }
 ts-rs = { version = "6.2", optional = true }
 unicase = "2.6.0"
 url = "2.3.1"
